Modified parser to handle incomplete messages
Made modifications to the parser portion of the JSONTools library to better handle
when a message is incompelte. It will now return the MESSAGE_INCOMPLETE status and
require a reset of the parser object. It will also read past a successful message to
see if there is another one in the buffer and alert the user to it.